get_implied_rl: allow for the possibility of an empty range.
In the code

 a = NULL;
 for (i=0; i<10;i++) {
       if (a) FOO;
       ...
 }

code that checks 'FOO' and examines the implied range of 'a' will
find that it has the full range rather than an empty range.
This is because handle_variable() returns NULL for an empty range
and _get_rl() treats this as failure.

Now that _get_rl() can successfully report an empty range,
handle_variable can as well.  A validation test is added to confirm
this works.

Signed-off-by: NeilBrown <neil@brown.name>